Fix home page nft & create NftUtils
